change script test setup
this prevents test timeout
