+++ server/test/mocha.opts
@@ -4,4 +4,4 @@
 ./dist/modes/template/test
 ./dist/modes/style/stylus/test
 ./dist/modes/test/
-./dist/service/
+./dist/modes/script/
